ppt-api/requirements.txt
python-pptx==0.6.21
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6 